go/tendermint/apps/staking: Move transaction handlers to its own file
